Rename host api files
The struct names were changed, so rename the files accordingly.
